+++ pyproject.toml
@@ -1,32 +1,27 @@
-[tool.poetry]
+[project]
 name = "ingest-wikimedia"
 version = "0.2.0"
-description = "Uploading media to Wikimedia Commons"
-authors = ["Scott Williams <scott@dp.la>, Michael Della Bitta <michael@dp.la>"]
-package-mode = false
+description = "Uploading media from DPLA to Wikimedia Commons"
+authors = [
+    { name = "Scott Williams", email = "scott@dp.la>" },
+    { name = "Michael Della Bitta", email = "michael@dp.la" },
+]
+license = { text = "MIT License" }
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "beautifulsoup4>=4.12.3",
+    "boto3-stubs[essential]>=1.35.49",
+    "boto3>=1.35.49",
+    "click>=8.1.7",
+    "python-magic>=0.4.27",
+    "pywikibot>=9.4.1",
+    "requests>=2.32.3",
+]
 
-[tool.poetry.dependencies]
-python = ">=3.12,<4.0"
-pywikibot = "^9.4.1"
-python-magic = "^0.4.27"
-requests = "^2.32.3"
-boto3 = "^1.35.46"
-click = "^8.1.7"
-boto3-stubs = {extras = ["essential"], version = "^1.35.40"}
-beautifulsoup4 = "^4.12.3"
-
-
-[tool.black]
-line-length = 88
-extend-exclude = "sdc-sync.py"
-
-[tool.poetry.group.dev.dependencies]
-black = "24.4.2"
-flake8 = "7.1.0"
-pytest = "^8.3.3"
-mypy = "^1.12.1"
-types-requests = "^2.32.0.20241016"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+[dependency-groups]
+dev = [
+    "mypy>=1.13.0",
+    "pytest>=8.3.3",
+    "types-requests>=2.32.0.20241016",
+]
