--- setup.cfg
@@ -1,3 +1,0 @@
-[flake8]
-max-line-length = 88
-extend-exclude = sdc-sync.py,user-config.py,user-password.py
